# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WeeklyRun API",
    "description": "WeeklyRun - AI Meeting Orchestration API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/filters/meetings/{meeting_id}/filtered": {
      "get": {
        "tags": [
          "filters"
        ],
        "summary": "Get Filtered Content",
        "description": "Get all filtered content for a meeting.\n\nArgs:\n    meeting_id: Meeting ID\n    include_restored: If True, include content that was restored (default: False)\n\nReturns:\n    List of filtered content items",
        "operationId": "get_filtered_content_api_v1_filters_meetings__meeting_id__filtered_get",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Meeting Id"
            }
          },
          {
            "name": "include_restored",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Restored"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilteredContentListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/filters/meetings/{meeting_id}/filtered/stats": {
      "get": {
        "tags": [
          "filters"
        ],
        "summary": "Get Filter Stats",
        "description": "Get statistics about filtered content for a meeting.",
        "operationId": "get_filter_stats_api_v1_filters_meetings__meeting_id__filtered_stats_get",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilterStatsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/filters/meetings/{meeting_id}/filtered/restore-all": {
      "post": {
        "tags": [
          "filters"
        ],
        "summary": "Restore All Filtered Content",
        "description": "Restore all filtered content for a meeting.",
        "operationId": "restore_all_filtered_content_api_v1_filters_meetings__meeting_id__filtered_restore_all_post",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Restore All Filtered Content Api V1 Filters Meetings  Meeting Id  Filtered Restore All Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/filters/meetings/{meeting_id}/filtered/confirm-all": {
      "post": {
        "tags": [
          "filters"
        ],
        "summary": "Confirm All Filtered Content",
        "description": "Confirm all non-restored filtered content as casual talk.",
        "operationId": "confirm_all_filtered_content_api_v1_filters_meetings__meeting_id__filtered_confirm_all_post",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Confirm All Filtered Content Api V1 Filters Meetings  Meeting Id  Filtered Confirm All Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/filters/meetings/{meeting_id}/filtered/{content_id}": {
      "get": {
        "tags": [
          "filters"
        ],
        "summary": "Get Filtered Content By Id",
        "description": "Get a specific filtered content item.",
        "operationId": "get_filtered_content_by_id_api_v1_filters_meetings__meeting_id__filtered__content_id__get",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Meeting Id"
            }
          },
          {
            "name": "content_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Content Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilteredContentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/filters/meetings/{meeting_id}/filtered/{content_id}/restore": {
      "post": {
        "tags": [
          "filters"
        ],
        "summary": "Restore Filtered Content",
        "description": "Restore filtered content back to the transcript.\n\nThis marks the content as restored, indicating the user\ndisagrees with the AI's classification.",
        "operationId": "restore_filtered_content_api_v1_filters_meetings__meeting_id__filtered__content_id__restore_post",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Meeting Id"
            }
          },
          {
            "name": "content_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Content Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestoreResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "filters"
        ],
        "summary": "Undo Restore",
        "description": "Undo a restore action - mark content as filtered again.",
        "operationId": "undo_restore_api_v1_filters_meetings__meeting_id__filtered__content_id__restore_delete",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Meeting Id"
            }
          },
          {
            "name": "content_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Content Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestoreResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/filters/meetings/{meeting_id}/filtered/{content_id}/confirm": {
      "post": {
        "tags": [
          "filters"
        ],
        "summary": "Confirm Filtered Content",
        "description": "Confirm that filtered content is indeed casual talk.\n\nThis is used for AI learning - confirmed items can be used\nto improve future filtering accuracy.",
        "operationId": "confirm_filtered_content_api_v1_filters_meetings__meeting_id__filtered__content_id__confirm_post",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Meeting Id"
            }
          },
          {
            "name": "content_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Content Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfirmResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "meeting_id": {
            "type": "string",
            "format": "uuid",
            "title": "Meeting Id"
          },
          "is_confirmed": {
            "type": "boolean",
            "title": "Is Confirmed"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "id",
          "meeting_id",
          "is_confirmed",
          "message"
        ],
        "title": "ConfirmResponse",
        "description": "Response after confirming content as casual talk."
      },
...
      "FilterStatsResponse": {
        "properties": {
          "meeting_id": {
            "type": "string",
            "format": "uuid",
            "title": "Meeting Id"
          },
          "total_filtered": {
            "type": "integer",
            "title": "Total Filtered"
          },
          "by_reason": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "By Reason"
          },
          "restored_count": {
            "type": "integer",
            "title": "Restored Count"
          },
          "confirmed_count": {
            "type": "integer",
            "title": "Confirmed Count"
          },
          "average_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Confidence"
          }
        },
        "type": "object",
        "required": [
          "meeting_id",
          "total_filtered",
          "by_reason",
          "restored_count",
          "confirmed_count",
          "average_confidence"
        ],
        "title": "FilterStatsResponse",
        "description": "Statistics about filtered content for a meeting."
      },
      "FilteredContentListResponse": {
        "properties": {
          "meeting_id": {
            "type": "string",
            "format": "uuid",
            "title": "Meeting Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/FilteredContentResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "restored_count": {
            "type": "integer",
            "title": "Restored Count"
          },
          "confirmed_count": {
            "type": "integer",
            "title": "Confirmed Count"
          }
        },
        "type": "object",
        "required": [
          "meeting_id",
          "items",
          "total_count",
          "restored_count",
          "confirmed_count"
        ],
        "title": "FilteredContentListResponse",
        "description": "List response for filtered contents."
      },
      "FilteredContentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "meeting_id": {
            "type": "string",
            "format": "uuid",
            "title": "Meeting Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "filter_reason": {
            "type": "string",
            "title": "Filter Reason"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "is_restored": {
            "type": "boolean",
            "title": "Is Restored"
          },
          "is_confirmed": {
            "type": "boolean",
            "title": "Is Confirmed"
          },
          "speaker_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speaker Label"
          },
          "speaker_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speaker Name"
          },
          "start_time": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "meeting_id",
          "content",
          "filter_reason",
          "confidence",
          "is_restored",
          "is_confirmed",
          "speaker_label",
          "speaker_name",
          "start_time",
          "end_time",
          "created_at"
        ],
        "title": "FilteredContentResponse",
        "description": "Filtered content response."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RestoreResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "meeting_id": {
            "type": "string",
            "format": "uuid",
            "title": "Meeting Id"
          },
          "is_restored": {
            "type": "boolean",
            "title": "Is Restored"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "id",
          "meeting_id",
          "is_restored",
          "message"
        ],
        "title": "RestoreResponse",
        "description": "Response after restoring filtered content."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}